+++ src/modules/dateTimePicker/components/index.ts
@@ -3,3 +3,4 @@ export * from './dateTimePicker/dateTimePicker.component';
 export * from './dateTimePicker/dateTimePicker.interface';
 export * from './dayPicker/dayPicker.component';
 export * from './monthPicker/monthPicker.component';
+export * from './yearPicker/yearPicker.component';